+++ CHANGES.md
@@ -1,8 +1,13 @@
 # Changes
 Recently completed and planned work is tracked here.
 
-## [0.0.0](.) [0.0.15](.)
+## [0.0.0](.) to [0.1.0](.)
 - Created early versions
 
-## [0.1.0](.)
-- Completed 1st release
+## [0.1.1](.)
+- Removed confusing `module_internal_version`
+- Removed obsolete `s3d-post-flow`
+- Updated `dev-setup`
+- Updated `name` module
+- Updated `sg_egress` module
+- Updated `sg_ingress` module